+++ packages/features/tasker/api/cleanup.ts
@@ -1,13 +1,14 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
 
-import tasker from "..";
+import { TaskProcessor } from "../task-processor";
 
 export async function GET(request: NextRequest) {
   const authHeader = request.headers.get("authorization");
   if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
     return new Response("Unauthorized", { status: 401 });
   }
-  await tasker.cleanup();
+  const processor = new TaskProcessor();
+  await processor.cleanup();
   return NextResponse.json({ success: true });
 }
