+++ packages/features/tasker/api/cron.ts
@@ -1,14 +1,15 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
 
-import tasker from "..";
+import { TaskProcessor } from "../task-processor";
 
 async function handler(request: NextRequest) {
   const authHeader = request.headers.get("authorization");
   if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
     return new Response("Unauthorized", { status: 401 });
   }
-  await tasker.processQueue();
+  const processor = new TaskProcessor();
+  await processor.processQueue();
   return NextResponse.json({ success: true });
 }
