+++ packages/features/tasker/tasker.ts
@@ -42,8 +42,6 @@ export type TaskerCreate = <TaskKey extends keyof TaskPayloads>(
 export interface Tasker {
   /** Create a new task with the given type and payload. */
   create: TaskerCreate;
-  processQueue(): Promise<void>;
-  cleanup(): Promise<void>;
   cancel(id: string): Promise<string>;
   cancelWithReference(referenceUid: string, type: TaskTypes): Promise<string | null>;
 }
